Change URIs of users and accounts
This revision removes the /user prefix and replaces it with /account.
This is because the username is already specified over HTTP basic
authentification, and therefore the username in the URI is unnecessary,
and would provide an attack window for hackers that try to trick the
server into giving them other users info by specifying the wrong URI.
Also, this makes the API better understandable.